+++ apps/machine/zynqmp_r5/CMakeLists.txt
@@ -1,4 +1,5 @@
 collect (APP_COMMON_SOURCES platform_info.c)
 collect (APP_COMMON_SOURCES rsc_table.c)
+collect (APP_COMMON_SOURCES zynqmp_r5_a53_rproc.c)
 collect (APP_INC_DIRS "${CMAKE_CURRENT_SOURCE_DIR}")
